PHP: After switching CLI to 5.3 in BOA-2.0.4 we should also symlink old 5.2 binary path to 5.3
This is to avoid confusion and to no longer use 5.2 on CLI just because it is still in the path and Drush or Aegir could call it anyway via `php` instead of a full path. Note that this is default in Drush 5 and Aegir 2.x vanilla to call just `php` and ignore full path, so we need to fix this now, before we will update our Aegir 2.x version to use Drush 5.