Use static downloads for all distros in stable edition.
Distros are usually build partially from the download on d.o, but often include extra contrib modules with dev releases, which could make BOA stable instances different, depending on the build or upgrade date. Furthermore, some distros are build from scratch with makefiles, also referencing contrib modules with dev releases and this is not only not so stable, but also extremely slow. We need to speed up things for stable BOA editions and make them as static as possible, where possible.