Run NewRelic re-install early enough to avoid locking full-upgrade in the waiting forever mode.
This may cause extended downtime during unattended upgrades when there is new version released and the new NewRelic packages may simply hang the process for some weird reason, like probably because of not honoring the non-interactive mode and waiting forever for y/n response. We are avoiding this by uninstalling NewRelic completely before running system full-upgrade and then doing this again if later in the process newer version of PHP happened to be installed.