Issue #2015553 - Fix the logic for protected registration of new accounts.
We no longer force anything live in the global.inc file (with exception of feature_server and hostmaster profiles). Instead, we use the running daily.sh maintenance script to detect actual variable per site and set it in the database directly, only if needed. This allows us to avoid opening moderated registration if the site already has "Administrators only" mode turned on. The logic for control files, as explained at http://omega8.cc/node/273 remains the same - we just use it in the running daily.sh script instead of check them live on every request. This allows to force safe defaults daily and to override them still in the site admin area (until the next daily.sh run).