Deprecate CiviCRM as a separate platform.
It no longer makes any sense to provide separate CiviCRM platform with new/latest version. CiviCRM is supported in any D7 or D6 platform and for reliable and consistent results it needs to live in the sites/all/modules space, at least until it will be supported by Aegir also in the profiles space, which could enable sharing the code again.