You can safely apply db updates via 'Run db updates' task on any site.
New task runs database updates on the site -- but first it will create a hot-backup to be able to revert the site database to the pre-update state in case there will be any problem detected. NOTE: This task is NOT a replacement for proper upgrade workflow and should be used only occasionally.