Set _PERMISSIONS_FIX=YES by default again, so critically important fixes can be applied.
It will still honor the fix_files_permissions_daily INI value set in the platform level INI file, but only if the platform has been already patched. If not, it will be ignored and BOA will proceed with patching the platform and forcing its default permissions/ownership on dirs/files in the codebase. Two weeks is more than enough and it is better to break the codebase git beauty than to leave it vulnerable.