PHP: Always install and use by default version 5.5
PHP 5.3 is no longer supported and shouldn't be used in production. We are switching to use PHP 5.5 by default, and it must be always installed, even if not specified in _PHP_MULTI_INSTALL, because it is used as a fallback option when PHP version is incorrectly specified in config files for any reason.