Make it possible to opt-out from otherwise recommended BOA Skynet auto-updates - fixes #557
Note that this is only first step to be included in the upcoming BOA-2.4.0 release, while BOA-2.4.1 should switch this to opt-in mode and disabled by default. It should be left for the system admin decision, but we also can't introduce opt-in mode without prior notice which should be published in the release notes.