Dev mode detection and URLs protection - works only for aliases
Using dev-like keywords in the main site name is no longer supported, because it is a bad and confusing practice. The dev-like mode should be used only for WSOD debugging and quick theme fixing to disable CSS/JS aggregation on the fly, but for nothing else, to avoid bad surprises while moving site into live mode.