Do not send extra nocache cookie on GET requests
Most of these requests are already protected by Drupal own cookie, or protected on the Nginx level with related exceptions. Sending this extra cookie on GET could results with allowing not logged in visitors to hit naked Drupal also on unrelated URIs (in the next 15s) which should be always cached and not served directly.