Use READ-COMMITTED isolation level by default
This can be a bit slower but should help significantly with issues Drupal is not able to fix by itself for transaction heavy apps. Note that for cluster it will work only for transactions within the same node context because this isolation level is not supprted across the cluster in the master-master mode.