BOA versions naming convention update
HEAD is officially the new Stable and named Latest, with -rel suffix for rolling release, while the old Stable is still available as Legacy via in-legacy and up-legacy, but shouldn't be used for anything beyond upgrade debugging. The in-stable and up-stable are now aliases for in-head and up-head, while cutting edge development branch will be soon available via in-dev and up-dev.