Add Drush 11 available as drush11-bin
Also remove confusing drush9 alias and instead use real versions in all commands. The default drush command still maps to drush8. The newer drush10-bin and drush11-bin map to system Drush 10 and 11 respectively, since they are not very useful on command line anyway. The drush-launcher is now available as drushlr command and should be used when local site drush is present if drush8 is not enough for certain operations and extensions in Drupal 8 and newer.