Force Aegir frontend and backend to use PHP 7.4 unless empty ~/static/control/unlock-aegir-php.info exists
This allows to use PHP 8.1 in fpm.info as default web version for all hosted sites without the need to list all sites in the multi-fpm.info and without affecting Aegir which is still not fully compatible with PHP 8. This new feature forces enabling multi-fpm.info if not enabled already.